[SPARK-57923][SQL][SDP][TESTS] Address review comments
- Fix class doc to describe the no-subquery form
- Hoist toExpectedContext helper into CommandSuiteBase to avoid duplication
- Drop redundant COMMENT test (covered by comparePlans test)
- Rename parse tests to "... correctly parsed" for consistency
- Also assert the second column's type in the column list test
