+++ sql/core/src/test/scala/org/apache/spark/sql/execution/command/v2/CreateStreamingTableParserSuite.scala
@@ -17,29 +17,19 @@
 
 package org.apache.spark.sql.execution.command.v2
 
-import org.apache.spark.QueryContext
 import org.apache.spark.sql.catalyst.analysis.UnresolvedIdentifier
 import org.apache.spark.sql.catalyst.parser.ParseException
 import org.apache.spark.sql.catalyst.plans.logical.{CreateStreamingTable, TableSpec}
 import org.apache.spark.sql.connector.expressions.{ClusterByTransform, FieldReference, IdentityTransform}
 import org.apache.spark.sql.execution.SparkSqlParser
 import org.apache.spark.sql.execution.command.v1.CommandSuiteBase
-import org.apache.spark.sql.types.IntegerType
+import org.apache.spark.sql.types.{IntegerType, StringType}
 
 /**
- * The class contains tests for the `CREATE STREAMING TABLE ... AS ...` command
+ * The class contains tests for the `CREATE STREAMING TABLE ...` command (without a subquery).
  */
 class CreateStreamingTableParserSuite extends CommandSuiteBase {
   protected lazy val parser = new SparkSqlParser()
-
-  /** Turn an actual QueryContext into an equivalent ExpectedContext to no-op the check. */
-  private def toExpectedContext(actual: QueryContext): ExpectedContext = ExpectedContext(
-    objectType = actual.objectType(),
-    objectName = actual.objectName(),
-    startIndex = actual.startIndex(),
-    stopIndex = actual.stopIndex(),
-    fragment = actual.fragment()
-  )
 
   test("CREATE STREAMING TABLE without subquery is parsed correctly") {
     comparePlans(
@@ -68,57 +58,51 @@ class CreateStreamingTableParserSuite extends CommandSuiteBase {
     )
   }
 
-  test("CREATE STREAMING TABLE - PARTITIONED BY is honored") {
+  test("CREATE STREAMING TABLE - PARTITIONED BY is correctly parsed") {
     val cmd = parser
       .parsePlan("CREATE STREAMING TABLE st PARTITIONED BY (a)")
       .asInstanceOf[CreateStreamingTable]
     assert(cmd.partitioning == Seq(IdentityTransform(FieldReference(Seq("a")))))
   }
 
-  test("CREATE STREAMING TABLE - COMMENT is honored") {
-    val cmd = parser
-      .parsePlan("CREATE STREAMING TABLE st COMMENT 'my streaming table'")
-      .asInstanceOf[CreateStreamingTable]
-    assert(cmd.tableSpec.comment == Some("my streaming table"))
-  }
-
-  test("CREATE STREAMING TABLE - TBLPROPERTIES are honored") {
+  test("CREATE STREAMING TABLE - TBLPROPERTIES are correctly parsed") {
     val cmd = parser
       .parsePlan("CREATE STREAMING TABLE st TBLPROPERTIES ('key' = 'value', 'num' = '1')")
       .asInstanceOf[CreateStreamingTable]
     assert(cmd.tableSpec.properties == Map("key" -> "value", "num" -> "1"))
   }
 
-  test("CREATE STREAMING TABLE - CLUSTER BY is honored") {
+  test("CREATE STREAMING TABLE - CLUSTER BY is correctly parsed") {
     val cmd = parser
       .parsePlan("CREATE STREAMING TABLE st CLUSTER BY (a)")
       .asInstanceOf[CreateStreamingTable]
     assert(cmd.partitioning == Seq(ClusterByTransform(Seq(FieldReference(Seq("a"))))))
   }
 
-  test("CREATE STREAMING TABLE - USING provider is honored") {
+  test("CREATE STREAMING TABLE - USING provider is correctly parsed") {
     val cmd = parser
       .parsePlan("CREATE STREAMING TABLE st USING parquet")
       .asInstanceOf[CreateStreamingTable]
     assert(cmd.tableSpec.provider == Some("parquet"))
   }
 
-  test("CREATE STREAMING TABLE - DEFAULT COLLATION is honored") {
+  test("CREATE STREAMING TABLE - DEFAULT COLLATION is correctly parsed") {
     val cmd = parser
       .parsePlan("CREATE STREAMING TABLE st DEFAULT COLLATION UTF8_LCASE")
       .asInstanceOf[CreateStreamingTable]
     assert(cmd.tableSpec.collation == Some("UTF8_LCASE"))
   }
 
-  test("CREATE STREAMING TABLE - column list is honored") {
+  test("CREATE STREAMING TABLE - column list is correctly parsed") {
     val cmd = parser
       .parsePlan("CREATE STREAMING TABLE st (id INT, name STRING)")
       .asInstanceOf[CreateStreamingTable]
     assert(cmd.columns.map(_.name) == Seq("id", "name"))
     assert(cmd.columns.head.dataType == IntegerType)
+    assert(cmd.columns(1).dataType == StringType)
   }
 
-  test("CREATE STREAMING TABLE - IF NOT EXISTS is honored") {
+  test("CREATE STREAMING TABLE - IF NOT EXISTS is correctly parsed") {
     Seq(
       ("CREATE STREAMING TABLE IF NOT EXISTS st", true),
       ("CREATE STREAMING TABLE st", false)
